Change the startupMessage for the 3 way handshake comunication
