+++ Assets/Materials/Colors/CharacterCollor.mat
@@ -109,7 +109,7 @@ Material:
     - _ZWrite: 1
     m_Colors:
     - _BaseColor: {r: 0, g: 0.8679245, b: 0.09605073, a: 1}
-    - _Color: {r: 0, g: 0.8679245, b: 0.09605069, a: 1}
+    - _Color: {r: 0, g: 0.8679245, b: 0.09605073, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.19999996, g: 0.19999996, b: 0.19999996, a: 1}
   m_BuildTextureStacks: []
