+++ Assets/ScriptHolders/ModelInfoManager.prefab
@@ -11,9 +11,9 @@ GameObject:
   - component: {fileID: 7691266782358625286}
   - component: {fileID: 726517977931092721}
   - component: {fileID: 8378137310830731447}
-  - component: {fileID: 6432659995187116412}
+  - component: {fileID: 8935272046410995955}
   m_Layer: 0
-  m_Name: StartupMessage
+  m_Name: ModelInfoManager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -58,7 +58,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c7445fe7907c9154198057fe7ebda33e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &6432659995187116412
+--- !u!114 &8935272046410995955
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -67,6 +67,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 3828634763652253311}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ba8d930d09cdf4340a680cb745110ee8, type: 3}
+  m_Script: {fileID: 11500000, guid: 185311bed9bc8d34daa889b86f88084d, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
