--- Runtime/BehavioursInterfaces/ISendStartupMessage.cs
@@ -1,9 +1,0 @@
-
-namespace ReupVirtualTwin.behaviourInterfaces
-{
-    public interface ISendStartupMessage
-    {
-        public string buildVersion { get; }
-        public void SendMessage();
-    }
-}
