+++ Runtime/DataModels/ModelInfoMessage.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba8d930d09cdf4340a680cb745110ee8
+guid: 5443df9aa48ed604fbf294ac37f3879f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
