+++ Runtime/DependencyInjectors/ModelInfoManagerDependencyInjector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e04041408f0632c4eb3f8c7048aad291
+guid: 185311bed9bc8d34daa889b86f88084d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
