+++ Runtime/ManagerInterfaces/IModelInfoManager.cs
@@ -1,0 +1,9 @@
+using ReupVirtualTwin.dataModels;
+
+namespace ReupVirtualTwin.managerInterfaces
+{
+    public interface IModelInfoManager
+    {
+        public WebMessage<ModelInfoMessage> ObtainModelInfoMessage();
+    }
+}
