+++ Runtime/ManagerInterfaces/IModelInfoManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6e7c6f0d18b85874397fe5c2bfbf997b
+guid: ece3cf7bfeca325458c3e39860555987
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
