+++ Tests/PlayMode/ModelInfoManagerTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f5cdea773bb58f44c98f11d7b6f8bf15
+guid: c1ed0ea5975b75f40bd542700b7186a9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
